requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
langchain==0.2.0
langchain-groq==0.1.5
langgraph==0.0.69
langchain-community==0.2.0
chromadb==0.5.0
python-dotenv==1.0.1
streamlit==1.35.0
requests==2.32.0
